refactor to use userprofile view
