+++ api/src/application/user/user-controller.js
@@ -7,6 +7,7 @@ const config = require('../../../config');
 const Boom = require('boom');
 const views = require('./user-views');
 const errorMessages = require('./user-error-messages');
+const userFormData = require('./user-form-data');
 
 // e.g. convert { foo.bar: 'baz' } to { foo: { bar: 'baz' }}
 const expandDotPaths = function(object) {
@@ -25,7 +26,6 @@ module.exports = (
   emailService,
   renderTemplate,
   clientService,
-  userFormData,
   oidcProvider,
   imageService
 ) => {
@@ -85,8 +85,8 @@ module.exports = (
         const pictureMIME = request.payload.picture.hapi.headers['content-type'];
 
         if (pictureMIME === 'image/jpeg' || pictureMIME === 'image/png') {
-          const filename = uuid();
-          const bucket = filename.substring(0, 2);
+          const uuid = uuid();
+          const bucket = uuid.substring(0, 2);
           const filename = await imageService.uploadImageStream(request.payload.picture, `pictures/${bucket}/${filename}`);
 
           profile = Object.assign(profile, payload, { picture: filename });
@@ -102,167 +102,10 @@ module.exports = (
         }
 
         reply.redirect(request.query.redirect_uri)
-      } else if (error) {
-        validationErrorMessages = getValidationMessages(error);
-        if (validationErrorMessages.picture || validationErrorMessages['hapi.headers.content-type']) {
-          validationErrorMessages.picture = ['Must be JPEG or PNG image less than 1MB'];
-        }
       }
 
-      const profile = user.get('profile');
-      const getValue = (field) => {
-        return (request.payload && request.payload[field]) || get(profile, field, '');
-      };
-      reply.view('user-profile', {
-        returnTo: request.query.redirect_uri,
-        title: 'User Profile',
-        fields: [
-          {
-            name: 'name',
-            label: 'Name',
-            type: 'text',
-            value: getValue('name'),
-            error: validationErrorMessages.name,
-          },
-          {
-            name: 'given_name',
-            label: 'Given Name',
-            type: 'text',
-            value: getValue('given_name'),
-            error: validationErrorMessages.given_name,
-          },
-          {
-            name: 'family_name',
-            label: 'Family Name',
-            type: 'text',
-            value: getValue('family_name'),
-            error: validationErrorMessages.family_name,
-          },
-          {
-            name: 'middle_name',
-            label: 'Middle Name',
-            type: 'text',
-            value: getValue('middle_name'),
-            error: validationErrorMessages.middle_name,
-          },
-          {
-            name: 'nickname',
-            label: 'Nickname',
-            type: 'text',
-            value: getValue('nickname'),
-            error: validationErrorMessages.nickname,
-          },
-          {
-            name: 'preferred_username',
-            label: 'Preferred Username',
-            type: 'text',
-            value: getValue('preferred_username'),
-            error: validationErrorMessages.preferred_username,
-          },
-          {
-            name: 'profile',
-            label: 'Profile',
-            type: 'text',
-            value: getValue('profile'),
-            error: validationErrorMessages.profile,
-          },
-          {
-            name: 'picture',
-            label: 'Picture',
-            isFile: true,
-            accept: 'image/jpeg, image/png',
-            value: getValue('picture'),
-            error: validationErrorMessages.picture,
-          },
-          {
-            name: 'website',
-            label: 'Website',
-            type: 'text',
-            value: getValue('website'),
-            error: validationErrorMessages.website,
-          },
-          {
-            name: 'gender',
-            label: 'Gender',
-            type: 'text',
-            value: getValue('gender'),
-            error: validationErrorMessages.gender,
-          },
-          {
-            name: 'birthdate',
-            label: 'Birthdate',
-            type: 'text',
-            value: getValue('birthdate'),
-            error: validationErrorMessages.birthdate,
-          },
-          {
-            name: 'zoneinfo',
-            label: 'Timezone',
-            isDropdown: true,
-            options: userFormData.timezones.map(name => ({
-              label: name,
-              value: name,
-              selected: getValue('zoneinfo') === name
-            })),
-            value: getValue('zoneinfo'),
-            error: validationErrorMessages.zoneinfo,
-          },
-          {
-            name: 'locale',
-            label: 'Locale',
-            isDropdown: true,
-            options: Object.keys(userFormData.locales).map((value) => ({
-              label: userFormData.locales[value],
-              value,
-              selected: getValue('locale') === value,
-            })),
-            value: getValue('locale'),
-            error: validationErrorMessages.locale,
-          },
-          {
-            name: 'phone_number',
-            label: 'Phone Number',
-            type: 'text',
-            value: getValue('phone_number'),
-            error: validationErrorMessages.phone_number,
-          },
-          {
-            name: 'address.street_address',
-            label: 'Street Address',
-            type: 'text',
-            value: getValue('address.street_address'),
-            error: validationErrorMessages['address.street_address'],
-          },
-          {
-            name: 'address.locality',
-            label: 'Locality',
-            type: 'text',
-            value: getValue('address.locality'),
-            error: validationErrorMessages['address.locality'],
-          },
-          {
-            name: 'address.region',
-            label: 'Region',
-            type: 'text',
-            value: getValue('address.region'),
-            error: validationErrorMessages['address.region'],
-          },
-          {
-            name: 'address.postal_code',
-            label: 'Postal Code',
-            type: 'text',
-            value: getValue('address.postal_code'),
-            error: validationErrorMessages['address.postal_code'],
-          },
-          {
-            name: 'address.country',
-            label: 'Country',
-            type: 'text',
-            value: getValue('address.country'),
-            error: validationErrorMessages['address.country'],
-          },
-        ]
-      });
+      const viewContext = views.userProfile(user, request, error);
+      reply.view('user-profile', viewContext);
     },
 
     getForgotPasswordForm: function(request, reply, source, error) {
@@ -366,7 +209,6 @@ module.exports['@require'] = [
   'email/email-service',
   'render-template',
   'client/client-service',
-  'user/user-form-data',
   'oidc-provider',
   'image/image-service',
 ];
